requirements.txt
python-telegram-bot==20.7
pydub==0.25.1
SpeechRecognition==3.10.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
pynput==1.7.6
python-dotenv==1.0.0
sounddevice==0.4.6
numpy>=1.26.0
scipy>=1.11.0
whisper==1.1.10
